refactor: intersection observer 선언 위치 변경
TestHome.tsx -> HomeVideoSlider.tsx
